better command queue flushing
Instead of flushing the driver and command queue in RenderPass, we now
do that in the framegraph.

We flush the command queue after each pass -- which reduces memory
pressure on the command queue and improves parallel execution
with the backend thread. We flush the driver only after the last
pass -- maybe we could improve this in the future as the framegraph
gain more knowledge about what's going on.